fix preset and datapack config loading errors
